update
added multiple sizes on a product

+++ Backend/Licenta/src/main/java/com/StefanSergiu/Licenta/controller/ProductController.java
@@ -1,26 +1,21 @@
 package com.StefanSergiu.Licenta.controller;
 
 import com.StefanSergiu.Licenta.config.BucketName;
-import com.StefanSergiu.Licenta.dto.product.CreateNewProductModel;
-import com.StefanSergiu.Licenta.dto.product.PlainProductDto;
-import com.StefanSergiu.Licenta.dto.product.ProductDto;
-import com.StefanSergiu.Licenta.dto.product.ProductRequestModel;
+import com.StefanSergiu.Licenta.dto.favorite.FavoriteDto;
+import com.StefanSergiu.Licenta.dto.product.*;
 import com.StefanSergiu.Licenta.entity.Favorite;
 import com.StefanSergiu.Licenta.entity.Product;
 import com.StefanSergiu.Licenta.entity.UserInfo;
-import com.StefanSergiu.Licenta.repository.FavoriteRepository;
+import com.StefanSergiu.Licenta.repository.ProductRepository;
 import com.StefanSergiu.Licenta.service.FavoriteService;
 import com.StefanSergiu.Licenta.service.FileStore;
 import com.StefanSergiu.Licenta.service.ProductService;
 import com.StefanSergiu.Licenta.service.UserService;
-import com.amazonaws.HttpMethod;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.GeneratePresignedUrlRequest;
-import jakarta.persistence.criteria.Predicate;
-import jakarta.transaction.Transactional;
+
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -31,16 +26,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.parameters.P;
-import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.net.URL;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
 
 import static org.apache.http.entity.ContentType.*;
 //@Autowired
@@ -60,6 +51,8 @@ public class ProductController {
 
     @Autowired
     FavoriteService favoriteService;
+    @Autowired
+    private ProductRepository productRepository;
 
     @PostMapping("/admin/add")
     public ResponseEntity<ProductDto> addProduct(@RequestBody final CreateNewProductModel createNewProductModel){
@@ -69,17 +62,18 @@ public class ProductController {
 
     @PreAuthorize("permitAll()")
     @GetMapping
-    public ResponseEntity<Page<ProductDto>> getproducts(@RequestParam(name = "name", required = false) String name,
-                                                        @RequestParam(name = "brands",required = false)String brands,
-                                                        @RequestParam(name = "genders", required = false)String genders,
-                                                        @RequestParam(name = "category_name",required = false)String category_name,
-                                                        @RequestParam(name = "price",required = false)Float price,
-                                                        @RequestParam(name ="minPrice", required = false) Float minPrice,
-                                                        @RequestParam(name = "maxPrice", required = false) Float maxPrice,
-                                                        @RequestParam(name = "type_name", required = false) String type_name,
-                                                        @RequestParam(name = "attributes", required = false) String attributesParam,
-                                                        @RequestParam(name="pageNumber", defaultValue = "0") int pageNumber,
-                                                        @RequestParam(name = "size", defaultValue = "12") int size){
+    public ResponseEntity<Page<ProductCardDto>> getproducts(@RequestParam(name = "name", required = false) String name,
+                                                            @RequestParam(name = "brands",required = false)String brands,
+                                                            @RequestParam(name = "genders", required = false)String genders,
+                                                            @RequestParam(name = "category_name",required = false)String category_name,
+                                                            @RequestParam(name = "price",required = false)Float price,
+                                                            @RequestParam(name ="minPrice", required = false) Float minPrice,
+                                                            @RequestParam(name = "maxPrice", required = false) Float maxPrice,
+                                                            @RequestParam(name = "type_name", required = false) String type_name,
+                                                            @RequestParam(name = "attributes", required = false) String attributesParam,
+                                                            @RequestParam(name="pageNumber", defaultValue = "0") int pageNumber,
+                                                            @RequestParam(name = "sizes", required = false) String sizes,
+                                                            @RequestParam(name = "size", defaultValue = "12") int size){
 
         // Parse attributes parameter into a Map<String, String>
         Map<String, String> attributes = new HashMap<>();
@@ -92,18 +86,29 @@ public class ProductController {
                 }
             }
         }
+        ProductRequestModel request = new ProductRequestModel(name,brands,category_name,genders,price, minPrice, maxPrice, type_name,attributes,sizes);
 
-        ProductRequestModel request = new ProductRequestModel(name,brands,category_name,genders,price, minPrice, maxPrice, type_name,attributes);
+        Pageable pageable = PageRequest.of(pageNumber, size, Sort.Direction.ASC,"name");
+
+        Page<ProductCardDto> productsPage = productService.getAllProducts(request, pageable);
+
+        CacheControl cacheControl = CacheControl.maxAge(20, TimeUnit.SECONDS);
+        return ResponseEntity.ok().cacheControl(cacheControl).body(productsPage);
+    }
+    @PreAuthorize("permitAll()")
+    @GetMapping("/{productName}")
+    public ResponseEntity<List<ProductDto>> getProduct(@PathVariable final String productName){
+
         //get logged in user
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String username= authentication.getName();
         UserInfo user = userService.getLoggedInUser(username);
-        //---
+        System.out.println(user);
+        ////
         Integer userId = null;
         if (user != null) {
             userId = user.getId();
         }
-        //---
         //favorite list by logged user. If user is null, then all productDto's are false on favorite
         List<Favorite> favorites;
         if(userId != null){
@@ -112,48 +117,35 @@ public class ProductController {
             favorites = null;
         }
 
-
-        Pageable pageable = PageRequest.of(pageNumber, size, Sort.Direction.ASC,"name");
-
-        Page<ProductDto> productsPage = productService.getAllProducts(request, pageable)
-                .map(product -> {
-                    ProductDto productDto = ProductDto.from(product);
-
-//            // download the image for every fetched product
-//            byte[] imageData = fileStore.download(product.getImagePath(), product.getImageFileName());
-//            //set the image to the dto
-//            productDto.setImage(imageData);
-                    //placeholder for image downloading
-                    byte[] placeholderByteArray = new byte[16992];
-                    productDto.setImage(placeholderByteArray);
-
-                    //set isFavorite to false by default
-                    productDto.setIsFavorite(false);
-                    //for every favorite entry in favorites list, set the isFavorite flag in Dto
-                    if(favorites != null){
-                        for(Favorite favorite: favorites) {
-                            if (favorite.getProduct().getId() == product.getId()) {
-                                productDto.setIsFavorite(true);
-                                break;
-                            }
-                        }
+        List<Product> products = productService.getProductByName(productName);
+        List<ProductDto> productDtoList = new ArrayList<>();
+        for (Product product : products) {
+            ProductDto productDto = ProductDto.from(product);
+            if(favorites != null){
+                for(Favorite fav : favorites){
+                    if(fav.getProduct().getId() == product.getId()){
+                        productDto.setIsFavorite(true);
+                        break;
                     }
-                    return productDto;
-                });
-
-        return ResponseEntity.ok(productsPage);
-    }
-
-    @GetMapping("/{productId}")
-    public ResponseEntity<ProductDto> getProduct(@PathVariable final Long productId){
-        Product product = productService.getProduct(productId);
-        return new ResponseEntity<>(new ProductDto().from(product),HttpStatus.OK);
+                }
+            }
+            productDtoList.add(productDto);
+        }
+        return new ResponseEntity<>(productDtoList,HttpStatus.OK);
     }
     @DeleteMapping(value ="/admin/{id}")
     public ResponseEntity<ProductDto> deleteProduct(@PathVariable final Long id){
-        Product product = productService.deleteProduct(id);
+        Product product = productService.getProduct(id);
+
+        if(productRepository.countByName(product.getName())>1) {
+            Product deletedProduct = productService.deleteProduct(id);
+            System.out.println("Am ajuns aici!");
+            return new ResponseEntity<>(ProductDto.from(deletedProduct),HttpStatus.OK);
+        }
+
+        Product deletedProduct = productService.deleteProduct(id);
         fileStore.deleteImage(product.getImagePath(), product.getImageFileName());
-        return new ResponseEntity<>(ProductDto.from(product),HttpStatus.OK);
+        return new ResponseEntity<>(ProductDto.from(deletedProduct),HttpStatus.OK);
     }
 
     @GetMapping(value = "{id}/image/download")
@@ -167,6 +159,15 @@ public class ProductController {
         Product product = productService.getProduct(id);
         if (product == null) {
             return new ResponseEntity<>("Product not found!", HttpStatus.NOT_FOUND);
+        }
+        // check if name already exists in database
+        //if present, don't upload the image and only set the path to the existing one
+        if(productRepository.countByName(product.getName())!=1){
+            List<Product> products = productRepository.findAllByName(product.getName());
+            //set the existing image to the product
+            productService.updateProductImage(id, products.get(0).getImagePath(),products.get(0).getImageFileName());
+            System.out.println("am ajuns aiciaea");
+            return ResponseEntity.status(HttpStatus.OK).body("Image added succesfully");
         }
 
         //Check if the file is an image
